+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Automate backend code assessment with a classical python toolkit (flake8,
   isort, black, pylint, bandit)
 - Add a dedicated OpenEdx oauth2 authentication backend (SSO)
+- Add a dedicated OpenEdx OpenID connect authentication backend (SSO)
 
 ### Changed